requirements.txt
streamlit==1.12.2
requests==2.31.0
openai==0.11.0
python-decouple==3.8
langchain==0.0.146
